pyproject.toml
[tool.poetry]
name = "connect"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
aiokafka = "^0.11.0"
python-dotenv = "^1.0.1"
python-environ = "^0.4.54"
pydantic-settings = "^2.4.0"
redis = "^5.0.8"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
